labextension: Rename LeftPanelWidget to LeftPanel
