+++ labextension/src/widgets/LeftPanel.tsx
