+++ labextension/src/widgets/VolumesPanel.tsx
@@ -18,7 +18,7 @@ import * as React from 'react';
 import { Button, Switch, Zoom } from '@material-ui/core';
 import AddIcon from '@material-ui/icons/Add';
 import DeleteIcon from '@material-ui/icons/Delete';
-import { IVolumeMetadata } from './LeftPanelWidget';
+import { IVolumeMetadata } from './LeftPanel';
 import { IRPCError, rokErrorTooltip } from '../lib/RPCUtils';
 import { Input } from '../components/Input';
 import { Select, ISelectOption } from '../components/Select';
